requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
executing==1.2.0
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
